+++ projects/aca-shared/src/lib/models/viewer.rules.ts
