+++ projects/aca-shared/src/lib/services/app.extension.service.spec.ts
@@ -24,8 +24,8 @@
  */
 
 import { TestBed } from '@angular/core/testing';
-import { AppTestingModule } from '../testing/app-testing.module';
-import { AppExtensionService } from './extension.service';
+import { AppTestingModule } from '../../../../../src/app/testing/app-testing.module';
+import { AppExtensionService } from './app.extension.service';
 import { Store, Action } from '@ngrx/store';
 import { AppStore } from '@alfresco/aca-shared/store';
 import {
