+++ src/app/components/context-menu/context-menu.component.ts
@@ -34,7 +34,6 @@ import {
   Inject
 } from '@angular/core';
 import { MatMenuTrigger } from '@angular/material/menu';
-import { AppExtensionService } from '../../extensions/extension.service';
 import { AppStore, getAppSelection } from '@alfresco/aca-shared/store';
 import { Store } from '@ngrx/store';
 import { Subject } from 'rxjs';
@@ -43,6 +42,7 @@ import { ContentActionRef } from '@alfresco/adf-extensions';
 import { ContextMenuOverlayRef } from './context-menu-overlay';
 import { CONTEXT_MENU_DIRECTION } from './direction.token';
 import { Directionality } from '@angular/cdk/bidi';
+import { AppExtensionService } from '@alfresco/aca-shared';
 
 @Component({
   selector: 'aca-context-menu',
