+++ src/app/components/current-user/current-user.component.ts
@@ -32,7 +32,7 @@ import {
   getUserProfile,
   getLanguagePickerState
 } from '@alfresco/aca-shared/store';
-import { AppExtensionService } from '../../extensions/extension.service';
+import { AppExtensionService } from '@alfresco/aca-shared';
 
 @Component({
   selector: 'aca-current-user',
