+++ src/app/components/info-drawer/metadata-tab/metadata-tab.component.ts
@@ -31,9 +31,8 @@ import {
   OnDestroy
 } from '@angular/core';
 import { MinimalNodeEntryEntity } from '@alfresco/js-api';
-import { NodePermissionService, isLocked } from '@alfresco/aca-shared';
+import { NodePermissionService, isLocked, AppExtensionService } from '@alfresco/aca-shared';
 import { AppStore, infoDrawerMetadataAspect } from '@alfresco/aca-shared/store';
-import { AppExtensionService } from '../../../extensions/extension.service';
 import { AppConfigService, NotificationService } from '@alfresco/adf-core';
 import { Observable, Subject } from 'rxjs';
 import { Store } from '@ngrx/store';
