+++ src/app/components/search/search-libraries-results/search-libraries-results.component.ts
@@ -29,10 +29,10 @@ import { BreakpointObserver, Breakpoints } from '@angular/cdk/layout';
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Params } from '@angular/router';
 import { Store } from '@ngrx/store';
-import { AppExtensionService } from '../../../extensions/extension.service';
 import { ContentManagementService } from '../../../services/content-management.service';
 import { PageComponent } from '../../page.component';
 import { SearchLibrariesQueryBuilderService } from './search-libraries-query-builder.service';
+import { AppExtensionService } from '@alfresco/aca-shared';
 
 @Component({
   selector: 'aca-search-results',
