+++ src/app/components/trashcan/trashcan.component.ts
@@ -29,9 +29,9 @@ import { BreakpointObserver, Breakpoints } from '@angular/cdk/layout';
 import { Component, OnInit } from '@angular/core';
 import { Store } from '@ngrx/store';
 import { Observable } from 'rxjs';
-import { AppExtensionService } from '../../extensions/extension.service';
 import { ContentManagementService } from '../../services/content-management.service';
 import { PageComponent } from '../page.component';
+import { AppExtensionService } from '@alfresco/aca-shared';
 
 @Component({
   templateUrl: './trashcan.component.html'
